@@ -28,37 +28,39 @@
     "baidu-aip-sdk": "^2.3.9",
     "electron-clipboard-extended": "^1.1.1",
     "electron-updater": "^4.0.6",
-    "element-ui": "^2.5.4",
+    "element-ui": "^2.12.0",
     "iohook": "^0.5.1",
     "isomorphic-fetch": "^2.2.1",
     "lodash": "^4.17.15",
-    "log4js": "^4.0.1",
     "md5": "^2.2.1",
     "robotjs": "^0.5.1",
     "sbd": "^1.0.15",
     "shortcut-capture": "^1.2.1",
     "translation.js": "^0.7.9",
-    "vue": "^2.5.21",
-    "vue-router": "^3.0.1",
+    "vue": "^2.6.10",
+    "vue-router": "^3.1.3",
     "vuedraggable": "^2.23.0",
-    "vuex": "^3.0.1"
+    "vuex": "^3.1.1"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.123",
+    "@types/lodash": "^4.14.138",
     "@types/md5": "^2.1.33",
     "@types/node": "^11.11.0",
     "@types/semver": "^6.0.0",
     "@types/yaml": "^1.0.2",
-    "@vue/cli-plugin-babel": "^3.2.0",
-    "@vue/cli-plugin-eslint": "^3.2.0",
+    "@typescript-eslint/eslint-plugin": "^2.1.0",
+    "@typescript-eslint/parser": "^2.1.0",
+    "@vue/cli-plugin-babel": "^3.11.0",
+    "@vue/cli-plugin-eslint": "^3.11.0",
     "@vue/cli-plugin-typescript": "^3.3.0",
     "@vue/cli-service": "^3.6.0",
     "@vue/eslint-config-prettier": "^4.0.1",
     "@vue/eslint-config-typescript": "^3.2.0",
+    "ajv": "^6.10.2",
     "babel-eslint": "^10.0.1",
     "electron": "^3.0.0",
     "element-theme-chalk": "^2.11.1",
-    "eslint": "^5.8.0",
+    "eslint": "^5.16.0",
     "eslint-plugin-typescript": "^0.14.0",
     "eslint-plugin-vue": "^5.0.0",
     "leancloud-storage": "^3.12.0",
@@ -66,8 +68,8 @@
     "typescript": "^3.3.3333",
     "typescript-eslint-parser": "^22.0.0",
     "valine": "^1.3.4",
-    "vue-cli-plugin-electron-builder": "^1.1.3",
-    "vue-template-compiler": "^2.5.21",
+    "vue-cli-plugin-electron-builder": "^1.4.0",
+    "vue-template-compiler": "^2.6.10",
     "yaml": "^1.5.0"
   },
   "cmake-js": {
