@@ -1,4 +1,4 @@
-const _ = require("lodash");
+import * as _ from "lodash";
 export const constants = {
   appName: "CopyTranslator",
   nickName: "Zouwu",
